HIVE-28558: Drop HCatalog download page from the website
The HCatalog download page (https://hive.apache.org/general/hcatalogdownloads/) is mostly there for historical reasons. It was probably useful back in 2013 to inform users about the merge of HCatalog in Hive but for the past 10 years we have been releasing HCatalog as part of Hive so anyone who is using that module does not need to visit that obsolete page.

Moreover, the presence of the HCatalog download page adds an additional level of indirection for users that want to download recent Hive releases.