HIVE-28687: Move Bylaws from documentation to community section and drop obsolete content
The community section in the website already has a Bylaws section but the latter points to the wiki that is now deprecated. After the wiki migration (HIVE-28166) the Bylaws content is under the Documentation section in the website.

There are also four pages holding old proposals to change the bylaws that were incorporated in the main content some time ago and they are now obsolete and redundant.