HIVE-28688: Move BecomingACommitter from documentation to community section (Stamatis Zampetakis reviewed by Butao Zhang)
The content of the BecomingACommitter is under the documentation section of the Website but it should be under the community section where there is already an existing link that currently points to the obsolete wiki.

The page also contains an image that is currently broken so we can fix it as well as part of this change.

Close apache/hive-site#27